Fix some 64-bit mingw compile issues.
Just one more issue to fix (alloca), but with a hack, QF compiles (no clue
yet if it works: wine doesn't seem to be an option at this stage)

+++ libs/gib/regex.c
@@ -2509,10 +2509,10 @@ typedef struct
   DEBUG_PRINT_COMPILED_PATTERN (bufp, pat, pend);			\
 									\
   /* Restore register info.  */						\
-  high_reg = (unsigned long) POP_FAILURE_ITEM ();				\
+  high_reg = (uintptr_t) POP_FAILURE_ITEM ();				\
   DEBUG_PRINT2 ("  Popping high active reg: %d\n", high_reg);		\
 									\
-  low_reg = (unsigned long) POP_FAILURE_ITEM ();				\
+  low_reg = (uintptr_t) POP_FAILURE_ITEM ();				\
   DEBUG_PRINT2 ("  Popping  low active reg: %d\n", low_reg);		\
 									\
   for (this_reg = high_reg; (unsigned) this_reg >= low_reg; this_reg--)	\
@@ -3225,8 +3225,8 @@ re_match_2 (bufp, string1, size1, string2, size2, pos, regs, stop)
   unsigned num_regs = bufp->re_nsub + 1;
 
   /* The currently active registers.  */
-  unsigned long lowest_active_reg = NO_LOWEST_ACTIVE_REG;
-  unsigned long highest_active_reg = NO_HIGHEST_ACTIVE_REG;
+  uintptr_t lowest_active_reg = NO_LOWEST_ACTIVE_REG;
+  uintptr_t highest_active_reg = NO_HIGHEST_ACTIVE_REG;
 
   /* Information on the contents of registers. These are pointers into
      the input strings; they record just what was matched (on this
